Move block handler to its own datasource
